requirements.txt
requests==2.21.0
blessings==1.7
dnspython==1.16.0
